Fix Windows incompatibility
Problem: This test doesn't run on Windows because it hardcodes the
`/tmp/` path, which isn't available on all platforms.

Solution: Use the temporary directory for whichever platform we're on,
add a test to ensure that it runs without errors, and then update the
`.travis.yml` so that we can be confident that this will continnue
working.